[REFACTOR] MappedSuperClass 제거 (#177)
