+++ src/main/java/com/debatetimer/dto/customize/request/CustomizeTableCreateRequest.java
@@ -1,8 +1,7 @@
 package com.debatetimer.dto.customize.request;
 
-import com.debatetimer.domain.TimeBoxes;
 import com.debatetimer.domain.customize.CustomizeTable;
-import com.debatetimer.domain.customize.CustomizeTimeBox;
+import com.debatetimer.domain.customize.CustomizeTimeBoxes;
 import com.debatetimer.domain.member.Member;
 import jakarta.validation.Valid;
 import java.util.List;
@@ -18,9 +17,9 @@ public record CustomizeTableCreateRequest(
         return info.toTable(member);
     }
 
-    public TimeBoxes<CustomizeTimeBox> toTimeBoxes(CustomizeTable customizeTable) {
+    public CustomizeTimeBoxes toTimeBoxes(CustomizeTable customizeTable) {
         return IntStream.range(0, table.size())
                 .mapToObj(i -> table.get(i).toTimeBox(customizeTable, i + 1))
-                .collect(Collectors.collectingAndThen(Collectors.toList(), TimeBoxes::new));
+                .collect(Collectors.collectingAndThen(Collectors.toList(), CustomizeTimeBoxes::new));
     }
 }
