+++ src/main/java/com/debatetimer/dto/customize/response/CustomizeTableResponse.java
@@ -1,24 +1,24 @@
 package com.debatetimer.dto.customize.response;
 
-import com.debatetimer.domain.TimeBoxes;
 import com.debatetimer.domain.customize.CustomizeTable;
 import com.debatetimer.domain.customize.CustomizeTimeBox;
+import com.debatetimer.domain.customize.CustomizeTimeBoxes;
 import java.util.List;
 
 public record CustomizeTableResponse(long id, CustomizeTableInfoResponse info, List<CustomizeTimeBoxResponse> table) {
 
     public CustomizeTableResponse(
             CustomizeTable customizeTable,
-            TimeBoxes<CustomizeTimeBox> timeBasedTimeBoxes
+            CustomizeTimeBoxes customizeTimeBoxes
     ) {
         this(
                 customizeTable.getId(),
                 new CustomizeTableInfoResponse(customizeTable),
-                toTimeBoxResponses(timeBasedTimeBoxes)
+                toTimeBoxResponses(customizeTimeBoxes)
         );
     }
 
-    private static List<CustomizeTimeBoxResponse> toTimeBoxResponses(TimeBoxes<CustomizeTimeBox> timeBoxes) {
+    private static List<CustomizeTimeBoxResponse> toTimeBoxResponses(CustomizeTimeBoxes timeBoxes) {
         List<CustomizeTimeBox> customizeTimeBoxes = timeBoxes.getTimeBoxes();
         return customizeTimeBoxes
                 .stream()
