+++ src/main/java/com/debatetimer/dto/member/TableResponses.java
@@ -1,6 +1,5 @@
 package com.debatetimer.dto.member;
 
-import com.debatetimer.domain.DebateTable;
 import com.debatetimer.domain.customize.CustomizeTable;
 import java.util.Comparator;
 import java.util.List;
@@ -8,8 +7,8 @@ import java.util.stream.Collectors;
 
 public record TableResponses(List<TableResponse> tables) {
 
-    private static final Comparator<DebateTable> DEBATE_TABLE_COMPARATOR = Comparator
-            .comparing(DebateTable::getUsedAt)
+    private static final Comparator<CustomizeTable> DEBATE_TABLE_COMPARATOR = Comparator
+            .comparing(CustomizeTable::getUsedAt)
             .reversed();
 
     public static TableResponses from(List<CustomizeTable> customizeTables) {
