+++ src/main/java/com/debatetimer/dto/member/TableType.java
@@ -2,7 +2,5 @@ package com.debatetimer.dto.member;
 
 public enum TableType {
 
-    PARLIAMENTARY,
-    TIME_BASED,
     CUSTOMIZE;
 }
