+++ src/test/java/com/debatetimer/domain/customize/CustomizeTableTest.java
@@ -3,11 +3,13 @@ package com.debatetimer.domain.customize;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.junit.jupiter.api.Assertions.assertAll;
 
 import com.debatetimer.domain.member.Member;
 import com.debatetimer.dto.member.TableType;
 import com.debatetimer.exception.custom.DTClientErrorException;
 import com.debatetimer.exception.errorcode.ClientErrorCode;
+import java.time.LocalDateTime;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -23,6 +25,37 @@ class CustomizeTableTest {
             CustomizeTable customizeTable = new CustomizeTable();
 
             assertThat(customizeTable.getType()).isEqualTo(TableType.CUSTOMIZE);
+        }
+    }
+
+    @Nested
+    class ValidateTableName {
+
+        @ValueSource(strings = {"a bc가다9", "가0나 다ab", "ㄱㄷㅇㄹ", "漢字", "にほんご", "vielfältig"})
+        @ParameterizedTest
+        void 테이블_이름은_이모지를_제외한_글자만_가능하다(String name) {
+            Member member = new Member("default@gmail.com");
+            assertThatCode(() -> new CustomizeTable(member, name, "agenda", true, true, "pros", "cons"))
+                    .doesNotThrowAnyException();
+        }
+
+        @ValueSource(strings = {"a😀bc가다9", "🐥", "🥦"})
+        @ParameterizedTest
+        void 테이블_이름에_이모지를_넣을_수_없다(String name) {
+            Member member = new Member("default@gmail.com");
+            assertThatThrownBy(() -> new CustomizeTable(member, name, "agenda", true, true, "pros", "cons"))
+                    .isInstanceOf(DTClientErrorException.class)
+                    .hasMessage(ClientErrorCode.INVALID_TABLE_NAME_FORM.getMessage());
+        }
+
+        @Test
+        void 테이블_이름은_정해진_길이_이내여야_한다() {
+            Member member = new Member("default@gmail.com");
+            String longTableName = "f".repeat(CustomizeTable.TABLE_NAME_MAX_LENGTH + 1);
+
+            assertThatThrownBy(() -> new CustomizeTable(member, longTableName, "agenda", true, true, "pros", "cons"))
+                    .isInstanceOf(DTClientErrorException.class)
+                    .hasMessage(ClientErrorCode.INVALID_TABLE_NAME_LENGTH.getMessage());
         }
     }
 
@@ -83,6 +116,59 @@ class CustomizeTableTest {
             assertThatThrownBy(() -> new CustomizeTable(member, "name", "agenda", true, true, "pros", consName))
                     .isInstanceOf(DTClientErrorException.class)
                     .hasMessage(ClientErrorCode.INVALID_TEAM_NAME_FORM.getMessage());
+        }
+    }
+
+    @Nested
+    class UpdateUsedAt {
+
+        @Test
+        void 테이블의_사용_시각을_업데이트한다() throws InterruptedException {
+            Member member = new Member("default@gmail.com");
+            CustomizeTable table = new CustomizeTable(member, "tableName", "agenda", true, true, "찬성", "반대");
+            LocalDateTime beforeUsedAt = table.getUsedAt();
+            Thread.sleep(1);
+
+            table.updateUsedAt();
+
+            assertThat(table.getUsedAt()).isAfter(beforeUsedAt);
+        }
+    }
+
+    @Nested
+    class Update {
+
+        @Test
+        void 테이블_정보를_업데이트_할_수_있다() {
+            Member member = new Member("default@gmail.com");
+            CustomizeTable table = new CustomizeTable(member, "tableName", "agenda", true, true, "pros", "cons");
+            CustomizeTable renewTable = new CustomizeTable(member, "newName", "newAgenda", false, false, "newPros",
+                    "newCons");
+
+            table.updateTable(renewTable);
+
+            assertAll(
+                    () -> assertThat(table.getName()).isEqualTo("newName"),
+                    () -> assertThat(table.getAgenda()).isEqualTo("newAgenda"),
+                    () -> assertThat(table.getProsTeamName()).isEqualTo("newPros"),
+                    () -> assertThat(table.getConsTeamName()).isEqualTo("newCons"),
+                    () -> assertThat(table.isWarningBell()).isFalse(),
+                    () -> assertThat(table.isFinishBell()).isFalse()
+            );
+        }
+
+        @Test
+        void 테이블_업데이트_할_때_사용_시간을_변경한다() throws InterruptedException {
+            Member member = new Member("default@gmail.com");
+            CustomizeTable table = new CustomizeTable(member, "tableName", "agenda", true, true, "pros", "cons");
+            CustomizeTable renewTable = new CustomizeTable(member, "newName", "newAgenda", false, false, "newPros",
+                    "newCons");
+            LocalDateTime beforeUsedAt = table.getUsedAt();
+            Thread.sleep(1);
+
+            table.updateTable(renewTable);
+
+            assertThat(table.getUsedAt()).isAfter(beforeUsedAt);
         }
     }
 }
